Extract LifecycleConcern from TimeAnalyticsConcern
Fix Rubocop Metrics/ModuleLength offense by extracting lifecycle-related
methods into a separate concern.
